+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Make WebAppInitData mockable [#24](https://github.com/khamidjon-khamidov/flutter_telegram_web_app/pull/24)
 * Added BiometricManager support
 * Added section_separator_color
+* Added `isVerticalSwipesEnabled`, `enableVerticalSwipes()`, `disableVerticalSwipes()` apis
 
 ## 0.1.4
 * Remove `tgWebAppStartParam` - doesn't exist in telegram script